integration: re-add docker-compose.yml
overloading the outer docker-compose.yml for both dev and test started
to feel more confusing than just keeping them separate.

namely, it's impossible to un-set 'build:' in an override file, so I
split that out into a docker-compose.override.yml file, symlinked under
hack/overrides/ to keep ./hack/dc working, but then having the tests
*not* use it, even though they use the docker-compose.yml right next to
it, and even though it's normally respected by 'docker-compose'
commands.
