+++ .gitignore
@@ -9,6 +9,7 @@ web/public/elm.min.js
 web/public/bundle.js
 web/public/*.bundle.js
 web/public/main.css
+docker-compose.override.yml
 *.coverprofile
 *.coverprofile.*
 *.test
