--- docker-compose.override.yml
@@ -1,1 +1,0 @@
-./hack/overrides/dev.yml
