--- hack/overrides/dev.yml
@@ -1,18 +1,0 @@
-version: '3'
-
-services:
-  db:
-    ports: [6543:5432]
-
-  web:
-    ports: [8080:8080]
-    volumes:
-      - .:/src
-
-  worker:
-    ports:
-    - 7777:7777
-    - 7788:7788
-    environment:
-      CONCOURSE_BIND_IP: 0.0.0.0
-      CONCOURSE_BAGGAGECLAIM_BIND_IP: 0.0.0.0
