+++ integration/docker-compose.yml
@@ -2,23 +2,22 @@ version: '3'
 
 services:
   db:
-    image: postgres:${POSTGRES_TAG:-latest}
+    image: ${TEST_POSTGRES_IMAGE:-postgres}
     shm_size: 1gb
-    ports: [6543:5432]
+    ports: [5432]
     environment:
       POSTGRES_DB: concourse
       POSTGRES_USER: dev
       POSTGRES_PASSWORD: dev
 
   web:
-    build: .
-    image: concourse/concourse:local
+    image: ${TEST_CONCOURSE_DEV_IMAGE:-concourse/concourse:local}
     command: web
     depends_on: [db]
-    ports: [8080:8080]
+    ports: [8080]
     volumes:
-    - .:/src
-    - "./hack/keys:/concourse-keys"
+    - ..:/src
+    - ../hack/keys:/concourse-keys
     environment:
       CONCOURSE_SESSION_SIGNING_KEY: /concourse-keys/session_signing_key
       CONCOURSE_TSA_AUTHORIZED_KEYS: /concourse-keys/authorized_worker_keys
@@ -32,19 +31,17 @@ services:
       CONCOURSE_EXTERNAL_URL: http://localhost:8080
       CONCOURSE_ADD_LOCAL_USER: test:test,guest:guest
       CONCOURSE_MAIN_TEAM_LOCAL_USER: test
-      CONCOURSE_CLUSTER_NAME: dev
+      CONCOURSE_CLUSTER_NAME: test
+      CONCOURSE_ENABLE_PIPELINE_INSTANCES: "true"
       CONCOURSE_ENABLE_ACROSS_STEP: "true"
 
   worker:
-    build: .
-    image: concourse/concourse:local
+    image: ${TEST_CONCOURSE_DEV_IMAGE:-concourse/concourse:local}
     command: worker
     privileged: true
     depends_on: [web]
-    ports:
-    - 7777:7777
-    - 7788:7788
-    volumes: ["./hack/keys:/concourse-keys"]
+    ports: [7777, 7788]
+    volumes: ["../hack/keys:/concourse-keys"]
     stop_signal: SIGUSR2
     environment:
       CONCOURSE_TSA_PUBLIC_KEY: /concourse-keys/tsa_host_key.pub
@@ -52,9 +49,6 @@ services:
 
       CONCOURSE_LOG_LEVEL: debug
       CONCOURSE_TSA_HOST: web:2222
-
-      CONCOURSE_BIND_IP: 0.0.0.0
-      CONCOURSE_BAGGAGECLAIM_BIND_IP: 0.0.0.0
 
       # avoid using loopbacks
       CONCOURSE_BAGGAGECLAIM_DRIVER: overlay
